Ensure correctly detect language variant at first load, and only load supported (i18n translated) language variant and failback to 2 characters language code if not supported.
e.g. if the top html tag is

    <html lang="EN-CA">

it will first test if "en-ca" is translated , then check if "en" is supported, and correct i18n translation will be used.
this behavior consistent with klaro.show() command.

change <html lang="XX-XX"> then call klaro.show()  was working, but auto detection was not. so this fix sync the behavior